feat(joox): rename to class name
